fix: resolve I18n load error in specs and move reporting logic
- Move `report_topic_file_change` from `Main` to `BJCHelpers` to improve modularity.
- Remove `main.rb` requirement from `build_tools_spec.rb` to prevent global I18n translation loading that caused TypeErrors during testing.
- Update specs to test reporting logic via a helper class instead of the main runner.
