+++ utilities/build-tools/bjc_helpers.rb
@@ -45,6 +45,16 @@ module BJCHelpers
     "?topic=#{topic}/#{unit_folder}&course=#{course}.html"
   end
 
+  def report_topic_file_change(topic_file_path, original_content)
+    return if File.read(topic_file_path) == original_content
+
+    relative_path = topic_file_path.delete_prefix("#{@rootDir}/")
+    puts <<~NOTICE
+      NOTICE: Build tools modified topic file: #{relative_path}
+              Its generated unit review section was updated. Review this file before committing.
+    NOTICE
+  end
+
   # Methods below here are only visible by calling BJCHelpers.X
   class << self
     # TODO: This needs to use a topic model to get the correct sequence.
